Specify user names, and re-use their browsers
This makes tests a) easier to read and
b) run faster (because we no longer need to
stat a browser for them for each scenario).

If you use the same user name between scenarios
then the existing browser is reused.
At the end, we close all user' browsers down.

Intrestingly, we still have some interesting failures
showing up, but I'm not really thinking abouw where the
app code is at the moment, so much as the testing boilerplate.